backend/requirements.txt
# FastAPI y servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Base de datos
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Validación y configuración
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Autenticación
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP y scraping
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Google APIs
google-api-python-client==2.114.0
google-cloud-vision==3.5.0

# OCR
pytesseract==0.3.10
Pillow==10.2.0

# Excel
openpyxl==3.1.2
pandas==2.1.4

# Utilidades
python-dateutil==2.8.2
aiofiles==23.2.1

# Tareas programadas
apscheduler==3.10.4
